refactor: organize utilities under `utils` dir
