+++ src/internal/utils/stringify.test.ts
