+++ src/internal/utils/stringify.ts
